+++ devenv.nix
@@ -33,4 +33,6 @@
     poetry.package = pkgs.poetry;
     venv.enable = true;
   };
+
+  # Note: I'm not using devenv's pre-commit support as isn't portable
 }