+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add pre-commit
+- Add failing probe for testing error handling
+- Add pre-commit hook
+
+### Fixed
+- Handle exceptions from probes more gracefully
 
 ## [0.2.0] - 2023-05-24
 ### Added
@@ -18,3 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release with basic functionality in place.
 
+
+[Unreleased]: https://github.com/micktwomey/pingertron/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/micktwomey/pingertron/compare/0.1.0...0.2.0
+[0.1.0]: https://github.com/micktwomey/pingertron/releases/tag/0.1.0